+++ src/types/build.ts
@@ -1,5 +1,8 @@
+import type { NodeVersion } from './helper-options';
+
 export interface BuildExecutorOptions {
 	input: string;
 	output: string;
 	external?: string[];
+	nodeVersion?: NodeVersion;
 }